fix couchapps/all_docs tests
