+++ corehq/couchapps/tests/test_all_docs.py
@@ -1,5 +1,6 @@
 from corehq.dbaccessors.couchapps.all_docs import \
-    get_all_doc_ids_for_domain_grouped_by_db, get_doc_count_by_type
+    get_all_doc_ids_for_domain_grouped_by_db, get_doc_count_by_type, \
+    delete_all_docs_by_doc_type
 from dimagi.utils.couch.database import get_db
 from django.test import TestCase
 
@@ -7,18 +8,22 @@ from django.test import TestCase
 class AllDocsTest(TestCase):
     @classmethod
     def setUpClass(cls):
+        cls.main_db = get_db(None)
+        cls.users_db = get_db('users')
+        delete_all_docs_by_doc_type(cls.main_db, ('Application', 'CommCareUser'))
+        delete_all_docs_by_doc_type(cls.users_db, ('Application', 'CommCareUser'))
         cls.domain = 'all-docs-domain'
         cls.main_db_doc = {'_id': 'main_db_doc', 'domain': cls.domain,
                            'doc_type': 'Application'}
         cls.users_db_doc = {'_id': 'users_db_doc', 'domain': cls.domain,
                             'doc_type': 'CommCareUser'}
-        get_db(None).save_doc(cls.main_db_doc)
-        get_db('users').save_doc(cls.users_db_doc)
+        cls.main_db.save_doc(cls.main_db_doc)
+        cls.users_db.save_doc(cls.users_db_doc)
 
     @classmethod
     def tearDownClass(cls):
-        get_db(None).delete_doc(cls.main_db_doc)
-        get_db('users').delete_doc(cls.users_db_doc)
+        cls.main_db.delete_doc(cls.main_db_doc)
+        cls.users_db.delete_doc(cls.users_db_doc)
 
     def test_get_all_doc_ids_for_domain_grouped_by_db(self):
         self.assertEqual(
